Core/Misc: some random clean-up

+++ src/server/game/Entities/Player/Player.cpp
@@ -23772,7 +23772,7 @@ void Player::InitRunes()
         m_runes->SetRuneState(i);
     }
 
-    for (uint32 i = 0; i < NUM_RUNE_TYPES; ++i)
+    for (uint8 i = 0; i < NUM_RUNE_TYPES; ++i)
         SetFloatValue(PLAYER_RUNE_REGEN_1 + i, 0.1f);
 }
